usb: ehci-hcd: Perform usb resume in ehci_register call
Inside ehci_deregister, the port is suspended, hence in ehci_register
call we need to resume the port (if it is suspended).
This fixes and issue when "usb stop" is called followed by "usb start"
if the reset is not asserted for EHCI-ctrl between the two calls.
Without resuming the port, the ehci-ctrl is in some bad state and
doesn't detect connected device. If reset is asserted, ports will not
longer remain in suspend state, hence we check if port is suspended

Signed-off-by: Pankaj Dev <pankaj.dev@st.com>
Change-Id: I049d91ecf80876339a8b022b80590f095923b3ad
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/303898
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
Domain-Review: Fabrice GASNIER <fabrice.gasnier@foss.st.com>