ARM: dts: stm32: Add a node for the FWU metadata device for stm32mp257f-dk-u-boot.dtsi
The FWU metadata structure is accessed through the driver model
interface. On the stm32mp257f dk and ev1 boards, the FWU metadata is
stored on the uSD card. Add the fwu-mdata node on the u-boot specific
dtsi file for accessing the metadata structure.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: Iea60ab5ae3b6b99cccd6a4d68b015ec5a97351a9
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/368865
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/372398
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>