stm32mp: bsec: add support for non-secure mirror
If Cortex-M33 is TDCID, it creates a mirror of the OTP fuses at the
beginning of SRAM1. This mirror is non-secure, so U-Boot can access it
directly. Update the BSEC driver for that. If a memory-region property
is found in bsec node then it is the non-secure OTP mirror.

Change-Id: I6dd5e02ff7e8dafa0a4fb8db7bd44102b098f0e8
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/416650
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>