ARM: dts: stm32: overwrite usbotg_hs node for stm32mp215f-dk-u-boot.dtsi
Due to DT synchronization with kernel DT, an error is returned by
typec_get_device_from_usb().

In u-boot there's no driver associated with "adc-usb-c-connector"
compatible. Thus there's no point in detecting the cable as it is
the power source of the board. In Linux it's used to know the power
source profile (500mA, 1.5A or 3A) but it isn't used in u-boot.

So remove "typec" and "port" nodes that are useless to remove the
"No UCLASS_USB_TYPEC for remote-endpoint" error.

Change-Id: I02efa3f00110581acb6a07470d8494cc70d539bf
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/459017
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>